# viva_api/api/spec/openapi_3_1_0_generated.yaml
openapi: 3.1.0
info:
  title: sms-api
  version: 0.9.155
paths:
...
  /viva/v1/environments/resolve:
    post:
      tags:
      - Viva Core
      - Environments
      summary: Which environment would a run with this need be given? (select; never
        builds)
      operationId: core-resolve-environment
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveEnvironmentRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentModel'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /compose/v1/curated/ecoli:
    post:
      tags:
      - Compose Curated
      summary: Run a v2ecoli whole-cell simulation via process-bigraph
      description: 'Run a v2ecoli whole-cell E. coli simulation.


        Unlike Copasi/Tellurium, v2ecoli does not require an SBML upload.

        The biological model is pre-computed in the ParCa cache and the

        55 biological processes are composed at runtime via process-bigraph.'
      operationId: compose-run-v2ecoli
      parameters:
      - name: duration
        in: query
        required: false
        schema:
          type: number
          description: Simulation duration in seconds.
          default: 60.0
          title: Duration
        description: Simulation duration in seconds.
      - name: seed
        in: query
        required: false
        schema:
          type: integer
          description: Random seed for stochastic processes.
          default: 0
          title: Seed
        description: Random seed for stochastic processes.
      - name: interval
        in: query
        required: false
        schema:
          type: number
          description: Execution interval (timestep) in seconds.
          default: 1.0
          title: Interval
        description: Execution interval (timestep) in seconds.
      - name: features
        in: query
        required: false
        schema:
          type: string
          description: JSON list of feature modules, e.g. '["ppgpp_regulation"]'
          default: '[]'
          title: Features
        description: JSON list of feature modules, e.g. '["ppgpp_regulation"]'
      - name: cache_dir
        in: query
        required: false
        schema:
          type: string
          description: Absolute path to pre-computed ParCa cache inside container.
          default: /out/cache
          title: Cache Dir
        description: Absolute path to pre-computed ParCa cache inside container.
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComposeSimulationExperiment'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ComposeSimulationExperiment:
      properties:
        simulation_database_id:
          type: integer
          title: Simulation Database Id
        simulator_database_id:
          type: integer
          title: Simulator Database Id
        last_updated:
          type: string
          title: Last Updated
        metadata:
          additionalProperties:
            type: string
          type: object
          title: Metadata
      type: object
      required:
      - simulation_database_id
      - simulator_database_id
      title: ComposeSimulationExperiment
...
    DependencyModel:
      properties:
        name:
          type: string
          title: Name
        source:
          type: string
          title: Source
          default: pypi
        version:
          type: string
          title: Version
          default: ''
      type: object
      required:
      - name
      title: DependencyModel
...
    EnvironmentModel:
      properties:
        image:
          type: string
          title: Image
        spec_hash:
          type: string
          title: Spec Hash
        image_digest:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Digest
      type: object
      required:
      - image
      - spec_hash
      title: EnvironmentModel
      description: 'What exists: the image a run would pull, and the two identities
        -- the request it answers

        (``spec_hash``) and what was built (``image_digest``; null until something
        records it).'
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ResolveEnvironmentRequest:
      properties:
        kind:
          type: string
          enum:
          - explicit
          - derived
          title: Kind
        key:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
        variant:
          type: string
          title: Variant
          default: ''
        dependencies:
          items:
            $ref: '#/components/schemas/DependencyModel'
          type: array
          title: Dependencies
      type: object
      required:
      - kind
      title: ResolveEnvironmentRequest
      description: 'What is needed. ``explicit``: the environment named ``key`` (a
        commit, or its own marked tag),

        optionally its ``variant`` image. ``derived``: whatever satisfies ``dependencies``
        -- an EMPTY list

        is a real request, for a composite that needs only the built-ins.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...